Move all methods supporting save_metadata_and_generate_alerts to utils as private methods
