+++ src/datadoc/frontend/callbacks/dataset.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import datetime
 import logging
-import re
 from typing import TYPE_CHECKING
 
 import arrow
@@ -15,8 +14,6 @@ from pydantic import ValidationError
 
 from datadoc import config
 from datadoc import state
-from datadoc.constants import CHECK_OBLIGATORY_METADATA_DATASET_MESSAGE
-from datadoc.constants import MISSING_METADATA_WARNING
 from datadoc.frontend.callbacks.utils import VALIDATION_ERROR
 from datadoc.frontend.callbacks.utils import MetadataInputTypes
 from datadoc.frontend.callbacks.utils import find_existing_language_string
@@ -41,9 +38,6 @@ from datadoc.frontend.fields.display_dataset import (
 )
 from datadoc.frontend.fields.display_dataset import (
     MULTIPLE_LANGUAGE_DATASET_IDENTIFIERS,
-)
-from datadoc.frontend.fields.display_dataset import (
-    OBLIGATORY_DATASET_METADATA_IDENTIFIERS_AND_DISPLAY_NAME,
 )
 from datadoc.frontend.fields.display_dataset import TIMEZONE_AWARE_METADATA_IDENTIFIERS
 from datadoc.frontend.fields.display_dataset import DatasetIdentifiers
@@ -292,31 +286,4 @@ def accept_dataset_metadata_date_input(
         error + no_error
         if dataset_identifier == DatasetIdentifiers.CONTAINS_DATA_FROM
         else no_error + error
-    )
-
-
-def _has_exact_word(word: str, text: str) -> bool:
-    """Return True if excact word matches text."""
-    return bool(re.search(rf"\b{word}\b", text))
-
-
-def dataset_control(error_message: str | None) -> dbc.Alert | None:
-    """Check obligatory metadata values for dataset.
-
-    Args:
-        error_message(str): A message generated by ObligatoryDatasetWarning containing names of fields missing value.
-    """
-    missing_metadata = [
-        f[1]
-        for f in OBLIGATORY_DATASET_METADATA_IDENTIFIERS_AND_DISPLAY_NAME
-        if (error_message and _has_exact_word(f[0], error_message))
-    ]
-    if not missing_metadata:
-        return None
-    return build_ssb_alert(
-        AlertTypes.WARNING,
-        MISSING_METADATA_WARNING,
-        CHECK_OBLIGATORY_METADATA_DATASET_MESSAGE,
-        None,
-        missing_metadata,
     )
