+++ src/datadoc/frontend/callbacks/utils.py
@@ -3,7 +3,9 @@
 from __future__ import annotations
 
 import datetime
+import json
 import logging
+import re
 import warnings
 from typing import TYPE_CHECKING
 from typing import TypeAlias
@@ -18,15 +20,25 @@ from dash import html
 
 from datadoc import config
 from datadoc import state
+from datadoc.constants import CHECK_OBLIGATORY_METADATA_DATASET_MESSAGE
+from datadoc.constants import CHECK_OBLIGATORY_METADATA_VARIABLES_MESSAGE
+from datadoc.constants import MISSING_METADATA_WARNING
 from datadoc.frontend.components.builders import AlertTypes
 from datadoc.frontend.components.builders import build_ssb_alert
 from datadoc.frontend.components.identifiers import ACCORDION_WRAPPER_ID
 from datadoc.frontend.components.identifiers import SECTION_WRAPPER_ID
 from datadoc.frontend.components.identifiers import VARIABLES_INFORMATION_ID
+from datadoc.frontend.fields.display_dataset import (
+    OBLIGATORY_DATASET_METADATA_IDENTIFIERS_AND_DISPLAY_NAME,
+)
+from datadoc.frontend.fields.display_variables import (
+    OBLIGATORY_VARIABLES_METADATA_IDENTIFIERS_AND_DISPLAY_NAME,
+)
 
 if TYPE_CHECKING:
     import pathlib
 
+    import dash_bootstrap_components as dbc
     import pydantic
     from cloudpathlib import CloudPath
 
@@ -244,6 +256,124 @@ def select_tabs(keydown: dict, current_tab: str) -> str:
     return current_tab
 
 
+def _has_exact_word(word: str, text: str) -> bool:
+    """Return True if excact word matches text."""
+    return bool(re.search(rf"\b{word}\b", text))
+
+
+def dataset_control(error_message: str | None) -> dbc.Alert | None:
+    """Check obligatory metadata values for dataset.
+
+    Args:
+        error_message(str): A message generated by ObligatoryDatasetWarning containing names of fields missing value.
+    """
+    missing_metadata = [
+        f[1]
+        for f in OBLIGATORY_DATASET_METADATA_IDENTIFIERS_AND_DISPLAY_NAME
+        if (error_message and _has_exact_word(f[0], error_message))
+    ]
+    if not missing_metadata:
+        return None
+    return build_ssb_alert(
+        AlertTypes.WARNING,
+        MISSING_METADATA_WARNING,
+        CHECK_OBLIGATORY_METADATA_DATASET_MESSAGE,
+        None,
+        missing_metadata,
+    )
+
+
+def _parse_error_message(message: str) -> list | None:
+    """Parse a string containing an error message into Python objects.
+
+    This function processes an error message by removing predefined text and
+    attempting to parse the remaining string into a list of Python objects.
+
+    Args:
+        message (str): The error message string to be parsed.
+
+    Returns:
+        A list of parsed objects if the string can be successfully parsed.
+        returns None if the string is empty after processing and a empty list
+        if the string cannot be parsed into JSON.
+    """
+    parsed_string = message.replace("Obligatory metadata is missing: ", "").strip()
+    parsed_string = parsed_string.replace("'", '"')
+    if not parsed_string:
+        return None
+    try:
+        # Attempt to parse the JSON string
+        return json.loads(parsed_string)
+    except json.JSONDecodeError:
+        logger.exception("Error parsing JSON {e}")
+        return []
+
+
+def _get_dict_by_key(
+    metadata_list: list[dict[str, list[str]]],
+    key: str,
+) -> dict[str, list[str]] | None:
+    """Return the first dictionary containing the specified key.
+
+    This function searches through a list of dictionaries and returns the first
+    dictionary that contains the specified key.
+
+    Args:
+        metadata_list: A list of dictionaries to search.
+        key: The key to search for in each dictionary.
+
+    Returns:
+        The first dictionary containing the specified key,
+        or None if no such dictionary is found.
+
+    """
+    return next((item for item in metadata_list if key in item), None)
+
+
+def variables_control(error_message: str | None, variables: list) -> dbc.Alert | None:
+    """Check obligatory metadata for variables and return an alert if any metadata is missing.
+
+    This function parses an error message to identify missing obligatory metadata
+    fields for variables. If missing metadata is found, it generates an alert.
+
+    Args:
+        error_message: A message generated by ObligatoryVariableWarning
+            containing the variable short name and a list of field names with missing values.
+        variables: list of datadoc variables
+
+    Returns:
+        An alert object if there are missing metadata fields, otherwise None.
+    """
+    missing_metadata: list = []
+    error_message_parsed = _parse_error_message(str(error_message))
+    # for variable in state.metadata.variables:
+    for variable in variables:
+        if error_message_parsed:
+            fields_by_variable = _get_dict_by_key(
+                error_message_parsed,
+                variable.short_name,
+            )
+            if fields_by_variable is not None:
+                missing_metadata_field = [
+                    f[1]
+                    for f in OBLIGATORY_VARIABLES_METADATA_IDENTIFIERS_AND_DISPLAY_NAME
+                    if error_message and f[0] in fields_by_variable[variable.short_name]
+                ]
+                missing_metadata_fields_to_string = ", ".join(missing_metadata_field)
+                missing_metadata.append(
+                    f"{variable.short_name}: {missing_metadata_fields_to_string}",
+                )
+    if not missing_metadata:
+        return None
+    return build_ssb_alert(
+        AlertTypes.WARNING,
+        MISSING_METADATA_WARNING,
+        CHECK_OBLIGATORY_METADATA_VARIABLES_MESSAGE,
+        None,
+        missing_metadata,
+    )
+
+
 def save_metadata_and_generate_alerts(metadata: Datadoc) -> list:
     """Save the metadata document to disk and check obligatory metadata.
 
@@ -251,9 +381,6 @@ def save_metadata_and_generate_alerts(metadata: Datadoc) -> list:
         List of alerts including obligatory metadata warnings if missing,
         and success alert if metadata is saved correctly.
     """
-    from datadoc.frontend.callbacks.dataset import dataset_control
-    from datadoc.frontend.callbacks.variables import variables_control
-
     missing_obligatory_dataset = ""
     missing_obligatory_variables = ""
