+++ src/datadoc/frontend/callbacks/variables.py
@@ -2,22 +2,17 @@
 
 from __future__ import annotations
 
-import json
 import logging
 from typing import TYPE_CHECKING
 
 from pydantic import ValidationError
 
 from datadoc import state
-from datadoc.constants import CHECK_OBLIGATORY_METADATA_VARIABLES_MESSAGE
-from datadoc.constants import MISSING_METADATA_WARNING
 from datadoc.frontend.callbacks.utils import MetadataInputTypes
 from datadoc.frontend.callbacks.utils import find_existing_language_string
 from datadoc.frontend.callbacks.utils import parse_and_validate_dates
-from datadoc.frontend.components.builders import AlertTypes
 from datadoc.frontend.components.builders import build_edit_section
 from datadoc.frontend.components.builders import build_ssb_accordion
-from datadoc.frontend.components.builders import build_ssb_alert
 from datadoc.frontend.components.builders import build_variables_machine_section
 from datadoc.frontend.constants import INVALID_DATE_ORDER
 from datadoc.frontend.constants import INVALID_VALUE
@@ -26,15 +21,11 @@ from datadoc.frontend.fields.display_variables import (
     MULTIPLE_LANGUAGE_VARIABLES_METADATA,
 )
 from datadoc.frontend.fields.display_variables import NON_EDITABLE_VARIABLES_METADATA
-from datadoc.frontend.fields.display_variables import (
-    OBLIGATORY_VARIABLES_METADATA_IDENTIFIERS_AND_DISPLAY_NAME,
-)
 from datadoc.frontend.fields.display_variables import VARIABLES_METADATA_LEFT
 from datadoc.frontend.fields.display_variables import VARIABLES_METADATA_RIGHT
 from datadoc.frontend.fields.display_variables import VariableIdentifiers
 
 if TYPE_CHECKING:
-    import dash_bootstrap_components as dbc
     from dapla_metadata.datasets import model
 
 
@@ -322,94 +313,3 @@ def set_variables_values_inherit_dataset_derived_date_values() -> None:
                 VariableIdentifiers.CONTAINS_DATA_UNTIL,
                 state.metadata.dataset.contains_data_until,
             )
-
-
-def _parse_error_message(message: str) -> list | None:
-    """Parse a string containing an error message into Python objects.
-
-    This function processes an error message by removing predefined text and
-    attempting to parse the remaining string into a list of Python objects.
-
-    Args:
-        message (str): The error message string to be parsed.
-
-    Returns:
-        A list of parsed objects if the string can be successfully parsed.
-        returns None if the string is empty after processing and a empty list
-        if the string cannot be parsed into JSON.
-    """
-    parsed_string = message.replace("Obligatory metadata is missing: ", "").strip()
-    parsed_string = parsed_string.replace("'", '"')
-    if not parsed_string:
-        return None
-    try:
-        # Attempt to parse the JSON string
-        return json.loads(parsed_string)
-    except json.JSONDecodeError:
-        logger.exception("Error parsing JSON {e}")
-        return []
-
-
-def _get_dict_by_key(
-    metadata_list: list[dict[str, list[str]]],
-    key: str,
-) -> dict[str, list[str]] | None:
-    """Return the first dictionary containing the specified key.
-
-    This function searches through a list of dictionaries and returns the first
-    dictionary that contains the specified key.
-
-    Args:
-        metadata_list: A list of dictionaries to search.
-        key: The key to search for in each dictionary.
-
-    Returns:
-        The first dictionary containing the specified key,
-        or None if no such dictionary is found.
-
-    """
-    return next((item for item in metadata_list if key in item), None)
-
-
-def variables_control(error_message: str | None, variables: list) -> dbc.Alert | None:
-    """Check obligatory metadata for variables and return an alert if any metadata is missing.
-
-    This function parses an error message to identify missing obligatory metadata
-    fields for variables. If missing metadata is found, it generates an alert.
-
-    Args:
-        error_message: A message generated by ObligatoryVariableWarning
-            containing the variable short name and a list of field names with missing values.
-        variables: list of datadoc variables
-
-    Returns:
-        An alert object if there are missing metadata fields, otherwise None.
-    """
-    missing_metadata: list = []
-    error_message_parsed = _parse_error_message(str(error_message))
-    # for variable in state.metadata.variables:
-    for variable in variables:
-        if error_message_parsed:
-            fields_by_variable = _get_dict_by_key(
-                error_message_parsed,
-                variable.short_name,
-            )
-            if fields_by_variable is not None:
-                missing_metadata_field = [
-                    f[1]
-                    for f in OBLIGATORY_VARIABLES_METADATA_IDENTIFIERS_AND_DISPLAY_NAME
-                    if error_message and f[0] in fields_by_variable[variable.short_name]
-                ]
-                missing_metadata_fields_to_string = ", ".join(missing_metadata_field)
-                missing_metadata.append(
-                    f"{variable.short_name}: {missing_metadata_fields_to_string}",
-                )
-    if not missing_metadata:
-        return None
-    return build_ssb_alert(
-        AlertTypes.WARNING,
-        MISSING_METADATA_WARNING,
-        CHECK_OBLIGATORY_METADATA_VARIABLES_MESSAGE,
-        None,
-        missing_metadata,
-    )
